fix payment methods.

+++ CommonServer/Services/DatabaseService.ts
@@ -465,7 +465,7 @@ class DatabaseService<TBaseModel extends BaseModel> {
     }
 
     public async create(createBy: CreateBy<TBaseModel>): Promise<TBaseModel> {
-        const onCreate: OnCreate<TBaseModel> = await this._onBeforeCreate(
+        const onCreate: OnCreate<TBaseModel> = createBy.props.ignoreHooks ? { createBy, carryForward: [] } : await this._onBeforeCreate(
             createBy
         );
 
@@ -519,13 +519,15 @@ class DatabaseService<TBaseModel extends BaseModel> {
 
         try {
             createBy.data = await this.getRepository().save(createBy.data);
-            createBy.data = await this.onCreateSuccess(
-                {
-                    createBy,
-                    carryForward,
-                },
-                createBy.data
-            );
+            if (!createBy.props.ignoreHooks) {
+                createBy.data = await this.onCreateSuccess(
+                    {
+                        createBy,
+                        carryForward,
+                    },
+                    createBy.data
+                );
+            }
             return createBy.data;
         } catch (error) {
             await this.onCreateError(error as Exception);
@@ -653,7 +655,7 @@ class DatabaseService<TBaseModel extends BaseModel> {
                 );
 
             findBy.query = checkReadPermissionType.query;
-            let count: number = 0; 
+            let count: number = 0;
 
             if (distinctOn) {
                 const queryBuilder: SelectQueryBuilder<TBaseModel> = this.getQueryBuilder(this.modelName)
@@ -668,14 +670,14 @@ class DatabaseService<TBaseModel extends BaseModel> {
                 count = await queryBuilder.getCount();
 
             } else {
-               count = await this.getRepository().count({
+                count = await this.getRepository().count({
                     where: findBy.query as any,
                     skip: (findBy.skip as PositiveNumber).toNumber(),
                     take: (findBy.limit as PositiveNumber).toNumber(),
                 });
             }
 
-        
+
             let countPositive: PositiveNumber = new PositiveNumber(count);
             countPositive = await this.onCountSuccess(countPositive);
             return countPositive;
@@ -697,7 +699,7 @@ class DatabaseService<TBaseModel extends BaseModel> {
 
     private async _deleteBy(deleteBy: DeleteBy<TBaseModel>): Promise<number> {
         try {
-            const onDelete: OnDelete<TBaseModel> = await this.onBeforeDelete(
+            const onDelete: OnDelete<TBaseModel> = deleteBy.props.ignoreHooks ? { deleteBy, carryForward: [] } : await this.onBeforeDelete(
                 deleteBy
             );
             const beforeDeleteBy: DeleteBy<TBaseModel> = onDelete.deleteBy;
@@ -716,7 +718,7 @@ class DatabaseService<TBaseModel extends BaseModel> {
                 limit: LIMIT_MAX,
                 populate: {},
                 select: {},
-                props: beforeDeleteBy.props,
+                props: { ...beforeDeleteBy.props, ignoreHooks: true },
             });
 
             await this._updateBy({
@@ -726,19 +728,24 @@ class DatabaseService<TBaseModel extends BaseModel> {
                 } as any,
                 props: {
                     isRoot: true,
+                    ignoreHooks: true
                 },
             });
 
             const numberOfDocsAffected: number =
                 (await this.getRepository().delete(beforeDeleteBy.query as any))
                     .affected || 0;
+            
+            
+            if (!deleteBy.props.ignoreHooks) {
 
-            await this.onDeleteSuccess(
-                { deleteBy, carryForward },
-                items.map((i: TBaseModel) => {
-                    return new ObjectID(i._id!);
-                })
-            );
+                await this.onDeleteSuccess(
+                    { deleteBy, carryForward },
+                    items.map((i: TBaseModel) => {
+                        return new ObjectID(i._id!);
+                    })
+                );
+            }
 
             return numberOfDocsAffected;
         } catch (error) {
@@ -762,7 +769,7 @@ class DatabaseService<TBaseModel extends BaseModel> {
                     createdAt: SortOrder.Descending,
                 };
             }
-            const onFind: OnFind<TBaseModel> = await this.onBeforeFind(findBy);
+            const onFind: OnFind<TBaseModel> = findBy.props.ignoreHooks ? { findBy, carryForward: [] } : await this.onBeforeFind(findBy);
             const onBeforeFind: FindBy<TBaseModel> = onFind.findBy;
             const carryForward: any = onFind.carryForward;
 
@@ -824,13 +831,14 @@ class DatabaseService<TBaseModel extends BaseModel> {
                 decryptedItems,
                 onBeforeFind
             );
-
-            decryptedItems = await (
-                await this.onFindSuccess(
-                    { findBy, carryForward },
-                    decryptedItems
-                )
-            ).carryForward;
+            if (!findBy.props.ignoreHooks) {
+                decryptedItems = await (
+                    await this.onFindSuccess(
+                        { findBy, carryForward },
+                        decryptedItems
+                    )
+                ).carryForward;
+            }
 
             return decryptedItems;
         } catch (error) {
@@ -934,7 +942,7 @@ class DatabaseService<TBaseModel extends BaseModel> {
 
     private async _updateBy(updateBy: UpdateBy<TBaseModel>): Promise<number> {
         try {
-            const onUpdate: OnUpdate<TBaseModel> = await this.onBeforeUpdate(
+            const onUpdate: OnUpdate<TBaseModel> = updateBy.props.ignoreHooks ? { updateBy, carryForward: [] } : await this.onBeforeUpdate(
                 updateBy
             );
 
@@ -961,7 +969,7 @@ class DatabaseService<TBaseModel extends BaseModel> {
                 limit: LIMIT_MAX,
                 populate: {},
                 select: {},
-                props: beforeUpdateBy.props,
+                props: { ...beforeUpdateBy.props, ignoreHooks: true },
             });
 
             for (let item of items) {
@@ -984,12 +992,14 @@ class DatabaseService<TBaseModel extends BaseModel> {
             //         )
             //     ).affected || 0;
 
-            await this.onUpdateSuccess(
-                { updateBy, carryForward },
-                items.map((i: TBaseModel) => {
-                    return new ObjectID(i._id!);
-                })
-            );
+            if (!updateBy.props.ignoreHooks) {
+                await this.onUpdateSuccess(
+                    { updateBy, carryForward },
+                    items.map((i: TBaseModel) => {
+                        return new ObjectID(i._id!);
+                    })
+                );
+            }
 
             return items.length;
         } catch (error) {
